requirements.txt
groq
python-dotenv
streamlit
pandas
plotly
requests
ics
pytest
plaid-python
flask